fixing concurrency problems and some refactor
